apps/worker/requirements.txt
redis==5.0.1
psycopg2-binary==2.9.9

